+++ praetorian_cli/handlers/chariot.py
@@ -6,5 +6,5 @@ from praetorian_cli.sdk.chariot import Chariot
 @click.group()
 @click.pass_context
 def chariot(ctx):
-    """ Chariot API access in the new and different file """
+    """ Command group for interacting with the Chariot product """
     ctx.obj = Chariot(keychain=ctx.obj)